Add files via upload
The files consisting of the dataset, the jupyter notebook of segmentation of the customers and the file app.py contains the dashboard based on the number of clusters(groups) of the customers. Here, the dashboard contains the scatter plot, and a table that contains the frequency and recency of the customers which gives information about the customer groups of their minimum purchasing.